AP-15347: Properly handle null value in `asString`.
Flow variables of type string can carry null values. The method
`asString` is intended to return a string representation of the value
(e.g. for display in the UI). As per this commit, if the value of the
flow variable is `null`, the string "null" is returned.

AP-15347 (Node Monitor fails to display values when a flow variable has null value)